+++ ChangeLog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - (Export) Combine multiple maps to single occlusion-roughness-metallic map
 - (Export) Emission support
 - (Export) Correct texture filter and wrap modes by creating glTF `sampler`
+- (Export) Support for injecting custom material conversion via `IMaterialExport`
 - (Documentation) XML documentation comments on many types
 - (Documentation) Initial setup for DocFX generator
 ### Changed